// domain/model/product-detail.entity.ts
export interface ItemDetailEntity {

  id: string
  site_id: string
  title: string
  subtitle: string
  seller_id: number
  category_id: string
  official_store_id: string
  price: number
  base_price: number
  original_price: number
  currency_id: string
  initial_quantity: number
  available_quantity: number
  sold_quantity: number
  sale_terms: string[]
  buying_mode: string
  listing_type_id: string
  start_time: string
  stop_time: string
  condition: string
  permalink: string
  thumbnail_id: string
  thumbnail: string
  secure_thumbnail: string
  pictures: string[]
  video_id: string
  descriptions: string[]
  accepts_mercadopago: boolean
  non_mercado_pago_payment_methods: string[]
  shipping: []
  international_delivery_mode: string
  seller_address: []
  seller_contact: string
  location: {}
  coverage_areas: []
  attributes: []
  warnings: []
  listing_source: string
  variations: []
  status: string
  sub_status: []
  tags: []
  warranty: string
  catalog_product_id: string
  domain_id: string
  parent_item_id: string
  differential_pricing: string
  deal_ids: []
  automatic_relist: boolean
  date_created: string
  last_updated: string
  health: string
  catalog_listing: boolean
  channels: []
  plain_text: string
}